Revert "Merge branch 'dev' into can_i_damageable-status-update"
This reverts commit e0f233b69fa476aa681caea9447269e32dcc9f65, reversing
changes made to feb211ed16d1dbe297db5446b71011117978fc48.

+++ src/game/CComponent.h
@@ -15,7 +15,7 @@ class CTextConsole;
 class CObjBase;
 
 
-enum COMP_TYPE : ushort
+enum COMP_TYPE : uchar
 {
     COMP_CHAMPION,
     COMP_SPAWN,
@@ -36,7 +36,7 @@ enum CCRET_TYPE
 class CComponent
 {
     COMP_TYPE _iType;
-
+    
 
 protected:
     CComponent(COMP_TYPE type);
